Implement some of the ULP Advisor's suggestions to reduce its whininess.

+++ radio.c
@@ -116,7 +116,7 @@ void init_radio() {
     write_single_register(0x25, 0b00000000); // GPIO map to default
 
     for (uint8_t sync_addr = 0x2f; sync_addr <= 0x36; sync_addr++) {
-        write_single_register(sync_addr, "QuCo12XII"[sync_addr%9]);
+        write_single_register(sync_addr, "QuCo12XII"[(uint8_t) (sync_addr%9)]);
     }
 
     // Setup addresses and length:
